Change whole layout, added in the two new files (see description)
I have added in the two new files which have been involved with the previous commit.  

I have also restyled the whole of the site.  Now there is a left hand side sidebar with a width rating of 20% on desktop & 60% on mobile devices.

The main content uses the calc function to figure out the size of itself.  

There is basic content within the sidebar area (this still needs work) and there is also basic content in the main area (this still needs work).